% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp_vig.R
\name{rcpp_vig}
\alias{rcpp_vig}
\title{PNAS journal format.}
\usage{
rcpp_vig(..., keep_tex = TRUE)
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Format for creating submissions to PNAS journals.
}
\examples{

\dontrun{
library(rmarkdown)
draft("MyArticle.Rmd", template = "pnas_article", package = "rticles")
}
}
