% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinp.R
\name{pinp}
\alias{pinp}
\title{PNAS-alike two-column journal format.}
\usage{
pinp(..., keep_tex = TRUE, citation_package = "natbib")
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{A boolean toggle to select whether intermediate
LaTeX files are to be kept, defaults to \code{TRUE}}

\item{citation_package}{A character variable selecting the LaTeX
citation package; defaults to \code{natbib}.}
}
\value{
R Markdown output format to pass to
\code{\link[rmarkdown:render]{render}}
}
\description{
Format suitable for attractive two-column pdf vignettes
}
\examples{
\dontrun{
library(rmarkdown)
draft("MyArticle.Rmd", template = "pinp", package = "pinp")
}
}
